% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_pi.R
\name{auc_pi}
\alias{auc_pi}
\title{Calculate area under the peak}
\usage{
auc_pi(bed = NULL, bw_files = NULL, bw_path = NULL, op_dir = NULL,
  numcores = NULL)
}
\arguments{
\item{bed}{A file in bed format. Default value is NULL.}

\item{bw_files}{One or a character vector with multiple files in bigWig format. Default value is NULL.}

\item{bw_path}{The path to directory where bwtool is installed on the computer. Default value is NULL.}

\item{op_dir}{The path to the operation directory currently used. Default value is NULL.}

\item{numcores}{Number of cores which should be used in parallelised process.Default value is NULL and will be defined as the number of available cores - 1.}
}
\description{
Takes files in bigWig format and one file in bed format and calculates the area under the curve (peak). Returns a data.table with the results.
}
