% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bw.R
\name{check_bw}
\alias{check_bw}
\title{checks for bwtool binary}
\usage{
check_bw(bw_path = NULL)
}
\arguments{
\item{bw_path}{The path to directory where bwtool is installed on the computer. Default value is NULL.}
}
\description{
Takes inserted path to bwtool and checks if bwtool is already installed.
}
