% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_avg.R
\name{get_avg}
\alias{get_avg}
\title{Order the entries in each matrix according to the average of rowMeans of all matrices combined.}
\usage{
get_avg(mat = NULL)
}
\arguments{
\item{mat}{A list with matrices and additional information about the selected region. mat generated by "get_matrix"-function. Default value is NULL.}
}
\value{
list with ordered matrix entries and additional information about the region entered in "get_matrix"-function. Will be used of "plot_hm"-function
}
\description{
Calculates average of rowMeans of all matrices in the list. Order the entries of all matrices according to this average. Called by "plot_hm"-function.
}
