% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix.R
\name{get_matrix}
\alias{get_matrix}
\title{Create a matrix with extracted data from a bigWig file within a specified region around a peak.}
\usage{
get_matrix(bed = NULL, bw_files = NULL, bw_path = NULL, op_dir = NULL,
  up = 2500, down = 2500, pos = "", binsize = 25)
}
\arguments{
\item{bed}{A file in bed format. Default value is NULL.}

\item{bw_files}{One or a character vector with multiple files in bigWig format. Default value is NULL.}

\item{bw_path}{The path to directory where bwtool is installed on the computer. Default value is NULL.}

\item{op_dir}{The path to the operation directory currently used. Default value is NULL.}

\item{up}{Number of basepairs from peak to 5' end. Default value is 2500.}

\item{down}{Number of basepairs from peak to 3' end.Default value is 2500.}

\item{pos}{Reference position of the region around a peak. Possibilities: '-starts' and '-ends'. Default value is '' and means a centered reference position.}

\item{binsize}{Binsize of how many basepairs the avergae will be calculated. Default value is 25.}
}
\value{
result list with matrices and additional information about the input of the function
}
\description{
Create a matrix with extracted data from bigWig files. Region around peak which should be observed can be specified. Returns a list with the matrix and the inserted parameters (region, binsize, reference position) and filenames (bed file andbigWig files).
}
