% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xavg.R
\name{get_xavg}
\alias{get_xavg}
\title{Order the entries in each matrix according to the rowMeans of the selected matrix (clusterBy = ).}
\usage{
get_xavg(mat = NULL, clusterBy = 1)
}
\arguments{
\item{mat}{A list with matrices and additional information about the selected region. mat generated by "get_matrix"-function. Default value is NULL.}

\item{clusterBy}{Integer which specifies the matrix to which all matrices will be ordered. Default value is 1.}
}
\value{
list with ordered matrix entries and additional information about the region entered in "get_matrix"-function. Will be used of "plot_hm"-function
}
\description{
Calculates rowMeans of selected matrix in the list. Order the entries of all matrices according to the ordered rowMeans. Called by "plot_hm"-function.
}
