\name{GRList-class}
\Rdversion{1.1}
\docType{class}
\alias{GRList-class}
\title{Class \code{"GRList"}}
\description{
Virutal Class to handle multiple data types for one slot
( \code{\link{profiles}})
}
\section{Objects from the Class}{A virtual Class:
No objects may be created from it.}
\section{Methods}{
    \describe{
    \item{\code{\link{GRList-class}}}{
    The purpose of this virtual classe is to store data of
    two different formats in  one slot: GRangesList and Lists}
    }
}
\references{
Zabet NR, Adryan B (2015) Estimating binding properties of transcription
factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94.

}
\author{
Patrick C. N. Martin <p.martin@essex.ac.uk>
}

\examples{
showClass("GRList")
}
\keyword{classes}
