% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Order the entries in each matrix according to the order returned by the "kmeans"-function of the selected reference matrix (refmat = ).}
\usage{
k_means(mat = NULL, refmat = NULL, k = NULL)
}
\arguments{
\item{mat}{A list with matrices and additional information about the selected region. mat generated by "get_matrix"-function. Default value is NULL.}

\item{refmat}{Integer which specifies the reference matrix to which all matrices will be ordered according to kmeans clustering of ference matrix. Default value is NULL.}

\item{k}{Integer which specifies the number of clusters which should be build by "kmeans"-function. Default value is NULL.}
}
\value{
list with ordered matrix entries, splitted matrices according to the clusters with column Means for profile plotting and additional information about the clusters and the region entered in "get_matrix"-function. Will be used of "plot_hm"-function
}
\description{
Clusters data of selected matrix in the list in k clusters. Order the entries of all matrices according to the ordered reference matrix. Called by "plot_hm"-function.
}
