% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bed.R
\name{make_bed}
\alias{make_bed}
\title{Create a temporary bed file.}
\usage{
make_bed(bed = NULL, op_dir = NULL)
}
\arguments{
\item{bed}{A file in bed format. Default value is NULL.}

\item{op_dir}{The path to the operation directory currently used. Default value is NULL.}
}
\description{
Create a temporary bed file for the following functions.
}
