\name{nos-class}
\Rdversion{1.1}
\docType{class}
\alias{nos-class}

\title{Class \code{"nos"}}
\description{
Virtual class to handle Number of Sites
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "nos" in the signature.
}
\references{
Zabet NR, Adryan B (2015) Estimating binding properties of transcription
factors from genome-wide binding profiles. Nucleic Acids Res., 43, 84–94.
Patrick C.N. Martin and Nicolae Radu Zabe (2020) Dissecting the binding mechanisms of transcription factors to DNA using a statistical thermodynamics framework. CSBJ, 18, 3590-3605.
}
\author{
Patrick C. N. Martin <pm16057@essex.ac.uk>

}
\examples{
showClass("nos")
}
\keyword{classes}
