% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_anno.R
\name{peak_anno}
\alias{peak_anno}
\title{Create Pie Chart with peak annotations}
\usage{
peak_anno(p_anno = NULL, state2color = NULL)
}
\arguments{
\item{p_anno}{File with peakID and corresponding Annotation. Default value is NULL.}
}
\description{
Creates a pie chart with peak annotations of entered file.
}
