% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annos.R
\name{peak_annos}
\alias{peak_annos}
\title{Create Pie Charts with peak annotations}
\usage{
peak_annos(p_anno = NULL, state2color = NULL)
}
\arguments{
\item{p_anno}{File with peakID and corresponding Annotation. Default value is NULL.}

\item{state2color}{File with 3 columns :1) region names as they are in the p_anno file; 2) region names as they should be shown in the image; 3) colors for plotting for each region name. Default value is NULL. If no state2color file is inserted function uses RColorBrewer palette}
}
\description{
Creates pie charts with peak annotations of entered files.
}
