% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hm.R
\name{plot_hm}
\alias{plot_hm}
\title{Plot a heatmap of ordered (and clustered) data in mat.}
\usage{
plot_hm(mat = NULL, clusterBy = "avg", num_k = NULL)
}
\arguments{
\item{mat}{A list with matrices and additional information about the selected region. mat generated by "get_matrix"-function. Default value is NULL.}

\item{clusterBy}{Integer which specifies the matrix to which all matrices will be ordered. Default value is 'avg'.}

\item{num_k}{Integer which specifies the number of clusters which should be build by "kmeans"-function. Default value is NULL.}
}
\description{
Takes entered list with matrices and passes it to "tidy_mat"-function and then to the functions "get_avg", "get_xavg" or "k_means" according to the entered values for clusterBy and k_num. Matrices will be ordered and then plotted with base graphics.
}
