% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MA.R
\name{plot_MA}
\alias{plot_MA}
\title{Plot the average of the condition group against the difference of the binary logarithm of the two conditions}
\usage{
plot_MA(aucpi = NULL, anno = NULL, cond = NULL)
}
\arguments{
\item{aucpi}{A data.table. Output of "auc_pi" function. Default value is NULL.}

\item{anno}{A data.frame with the assignment of the bigWig files to the conditions. Default value is NULL.}

\item{cond}{A character vector with two conditions according to the anno data.frame and the bigWig files which should be compared.}
}
\description{
Compare the mean of area under the curve signal for data sets according to the inserted conditions. The conditions are adjusted to the anno-file where the conditions for all data sets are collated. Plots the average of both inserted conditions against the difference of the binary logarithm of the two conditions.
}
