% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pol2i.R
\name{plot_pol2i}
\alias{plot_pol2i}
\title{Plot calculated polymerase2 pausing index.}
\usage{
plot_pol2i(pol2_table = NULL)
}
\arguments{
\item{pol2_table}{List generated with the "pol2_index"-function. Default value is NULL.}
}
\description{
Plot calculated polymerase2 pausing index for every gene according to the inserted gene model file in "pol2_index"-function.  With used filter duplicated and overlapping genes can be removed from list as well as non necessary columns. Returns a list (name, geneName, txName, chrom, strand, txStart, txEnd, nr, dr, pi, dif).
}
