% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot mean or median values of}
\usage{
plot_profile(mat_sum = NULL, opt = "mean")
}
\arguments{
\item{opt}{Option how the values of mat_sum should be summarized. Possible options are 'mean' or 'median'. Default value is mean.}

\item{mar_sum}{A list A list with average values of region around a peak. Output of "get_matrix" function. "plot_profile" function will take it as input to run "summarize" function. Default value is NULL.}
}
\value{
nothing.
}
\description{
Plot mean or median value for every inserted bigWig file for every region according to the adjustments made in "get_matrix" function.
}
