% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pol2_index.R
\name{pol2_index}
\alias{pol2_index}
\title{Calculate polymerase2 pausing index for every gene according to the inserted gene model file.}
\usage{
pol2_index(gm = NULL, bw_files = NULL, bw_path = NULL, op_dir = NULL,
  filter = TRUE, numcores = NULL)
}
\arguments{
\item{gm}{Gene model. File with information about the transcription start and end position of every gene. Default value is NULL.}

\item{bw_files}{One or a character vector with multiple files in bigWig format. Default value is NULL.}

\item{bw_path}{The path to directory where bwtool is installed on the computer. Default value is NULL.}

\item{op_dir}{The path to the operation directory currently used. Default value is NULL.}

\item{filter}{Filter which can be used to filter the output list to remove  duplicated and overlapping genes as well as non necessary columns. Default value is TRUE.}

\item{numcores}{Number of cores which should be used in parallelised process.Default value is NULL and will be defined as the number of available cores - 1.}
}
\value{
list (name, size_tss, sum_tss, size_gb, sum_gb, geneNAme, TxName, chrom ,strand, txStart, txEnd, nr, dr, pi). IF filter = TRUE : list (name, geneName, txName, chrom, strand, txStart, txEnd, nr, dr, pi, dif)
}
\description{
Calculate polymerase2 pausing index for every gene according to the inserted gene model file. Returns a list (name, size_tss, sum_tss, size_gb, sum_gb, geneNAme, TxName, chrom ,strand, txStart, txEnd, nr, dr, pi). With used filter duplicated and overlapping genes can be removed from list as well as non necessary columns. Returns a list (name, geneName, txName, chrom, strand, txStart, txEnd, nr, dr, pi, dif).
}
