% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Calculates mean or median for every region in a bed file.}
\usage{
summarize(mat_list, opt = "mean")
}
\arguments{
\item{mat_list}{A list with average values of region around a peak. Output of "get_matrix" function. Default value is NULL.}

\item{opt}{Option how the values of mat_list should be summarized. Possible options are 'mean' or 'median'. Default value is mean.}
}
\description{
Calculates mean or median for every region in a bed file. Takes output of "get_matrix" function as mat_list. Returns list for every inserted bigWig file with the mean or median value for every region according to the adjustments made in "get_matrix" function.
}
