% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annos.R
\name{peak_annos}
\alias{peak_annos}
\title{Create Pie Charts with peak annotations}
\usage{
peak_annos(p_anno = NULL, state2color = NULL)
}
\arguments{
\item{p_anno}{File with peakID and corresponding Annotation. Default value is NULL.}
}
\description{
Creates pie charts with peak annotations of entered files.
}
